Mindustry / Goliath Ship
This is my first working mod, i actually just kind of copy paste some stuff, cause i wasnt able to figure out why the mod would not work when i change the JSON-File names. I even named the sprites and everything correctly, but nevertheless, the mod is working. The names are just confusing, but the mod works fine and is editable trough the JSON-files. ENJOY. Complete folders are on the way.